+++ Koronavilkku/Koronavilkku/Tools/BackgroundTask.swift
@@ -84,7 +84,11 @@ final class BackgroundTaskForNotifications: BackgroundTask {
         Log.d("Register background task for notifications")
         BGTaskScheduler.shared.register(forTaskWithIdentifier: identifier, using: .main) { task in
             Log.d("Run background task for notifications")
-
+            
+            guard !Environment.default.exposureRepository.isEndOfLife else {
+                return task.setTaskCompleted(success: true)
+            }
+            
             // reschedule first to prevent unexpected errors from breaking the chain
             self.schedule()
 
@@ -207,6 +211,10 @@ fileprivate final class BackgroundTaskForDummyPosting: BackgroundTask {
         BGTaskScheduler.shared.register(forTaskWithIdentifier: identifier, using: .main) { task in
             Log.d("Run background task for dummy posting")
 
+            guard !Environment.default.exposureRepository.isEndOfLife else {
+                return task.setTaskCompleted(success: true)
+            }
+            
             // reschedule first to prevent unexpected errors from breaking the chain
             self.schedule()